Login page: migrate to TypeScript and CSS Modules
