--- src/pages/Login/Login.css
@@ -1,5 +1,0 @@
-.animate-list-accounts {
-  overflow: hidden;
-  transform: translateX(-200%);
-  transition: transform 1s ease-in-out;
-}
