+++ src/pages/Login/SetAccountPassword.module.css
@@ -1,0 +1,3 @@
+.pageWrapper {
+  background: rgb(var(--color-white));
+}
